Movement based off of openPaths array
Simplified moving a lot by just using the openPaths array and always moving to the shortest hCost according to the array. When the checker starts to wander away from the shortest path, it moves back on track because there are other openPaths with lower hCosts than the one it's currently on.